+++ CHANGELOG.markdown
@@ -10,6 +10,7 @@
 * Moved `(:<->)` to `Control.Lens.Simple`.
 * Fixed a bug in `Control.Lens.TH` that was causing `makeIso` not to work.
 * Added `lifted` to `Control.Lens.Setter` for mapping over monads.
+* Added `beside` to `Control.Lens.Traversal`.
 
 3.2
 ---